requirements.txt
streamlit
qdrant-client
langchain-cohere
langchain-community-vectorstores
langchain-groq
langchain-core
